4x bigger max. zoom in 2d mode at Gambini's request
git-svn-id: https://svn.eduke32.com/eduke32@1491 1a8010ca-5511-0410-912e-c29ae57300e0

+++ polymer/eduke32/source/mapster32.h
@@ -106,7 +106,7 @@ static inline void SetSLIMEPalette();
 static inline void SetWATERPalette();
 static inline void SetGAMEPalette();
 
-extern int16_t grid;
+extern int32_t grid;
 
 static void EditSpriteData(int16_t spritenum);
 static void EditWallData(int16_t wallnum);
